better handling of add column
prevent crashing when upgrading from blank db

+++ src/airunner/alembic/versions/8ebcece37db8_add_lora_scale_to_generator_settings.py
@@ -1,10 +1,3 @@
-"""Add lora_scale to generator_settings
-
-Revision ID: 8ebcece37db8
-Revises: 4626ae0d0601
-Create Date: 2024-10-04 10:16:43.172811
-
-"""
 from typing import Sequence, Union
 
 from alembic import op
@@ -17,10 +10,21 @@ down_revision: Union[str, None] = '4626ae0d0601'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
 
-
 def upgrade():
-    op.add_column('generator_settings', sa.Column('lora_scale', sa.Integer, default=100))
+    # Check if the 'lora_scale' column already exists
+    conn = op.get_bind()
+    inspector = sa.inspect(conn)
+    columns = [column['name'] for column in inspector.get_columns('generator_settings')]
+
+    if 'lora_scale' not in columns:
+        op.add_column('generator_settings', sa.Column('lora_scale', sa.Integer, default=100))
 
 def downgrade():
-    op.drop_column('generator_settings', 'lora_scale')
+    # Drop the 'lora_scale' column if it exists
+    conn = op.get_bind()
+    inspector = sa.inspect(conn)
+    columns = [column['name'] for column in inspector.get_columns('generator_settings')]
+
+    if 'lora_scale' in columns:
+        op.drop_column('generator_settings', 'lora_scale')
     # ### end Alembic commands ###
